+++ examples/progress-plots/liquid/demo/diamond.ts
@@ -1,0 +1,15 @@
+import { Liquid } from '@antv/g2plot';
+
+const liquidPlot = new Liquid('container', {
+  percent: 0.25,
+  shape: 'diamond',
+  outline: {
+    border: 4,
+    distance: 8,
+  },
+  wave: {
+    length: 128,
+  },
+});
+
+liquidPlot.render();