+++ src/plots/sankey/types.ts
@@ -1,4 +1,5 @@
-import { Data, Datum, Options, StyleAttr } from '../../types';
+import { Data, Options, StyleAttr } from '../../types';
+import { NodeDepth, NodeSort } from './layout';
 
 /** 配置类型定义 */
 export interface SankeyOptions extends Omit<Options, 'xField' | 'yField' | 'xAxis' | 'yAxis'> {
@@ -41,7 +42,11 @@ export interface SankeyOptions extends Omit<Options, 'xField' | 'yField' | 'xAxi
   /**
    * 节点排序方式，默认为空
    */
-  readonly nodeSort?: (a: Datum, b: Datum) => number;
+  readonly nodeSort?: NodeSort;
+  /**
+   * 节点排放分层的顺序，从 0 开始，并且返回值需要保证所有的层级都有节点
+   */
+  readonly nodeDepth?: NodeDepth;
   /**
    * 节点样式
    */