+++ src/plots/treemap/index.ts
@@ -2,6 +2,7 @@ import { Plot } from '../../core/plot';
 import { Adaptor } from '../../core/adaptor';
 import { TreemapOptions } from './types';
 import { adaptor } from './adaptor';
+import { transformData, isDrillDown } from './utils';
 import './interactions/treemap-drill-down';
 import '../scatter/interaction';
 
@@ -10,6 +11,21 @@ export { TreemapOptions };
 export class Treemap extends Plot<TreemapOptions> {
   /** 图表类型 */
   public type: string = 'treemap';
+
+  /**
+   * changeData
+   */
+  public changeData(data) {
+    const { colorField, interactions, hierarchyConfig } = this.options;
+    this.updateOption({ data });
+    const transData = transformData({
+      data,
+      colorField,
+      openDrillDown: isDrillDown(interactions),
+      hierarchyConfig,
+    });
+    this.chart.changeData(transData);
+  }
 
   protected getSchemaAdaptor(): Adaptor<TreemapOptions> {
     return adaptor;