+++ src/utils/hierarchy/treemap.ts
@@ -19,7 +19,13 @@ const DEFAULT_OPTIONS: HierarchyOption = {
   as: ['x', 'y'],
   // 默认降序
   sort: (a, b) => b.value - a.value,
+  // 纵横比, treemapSquarify 布局时可用，默认黄金分割比例
+  ratio: 0.5 * (1 + Math.sqrt(5)),
 };
+
+export function getTileMethod(tile: string, ratio: number) {
+  return tile === 'treemapSquarify' ? d3Hierarchy[tile].ratio(ratio) : d3Hierarchy[tile];
+}
 
 export function treemap(data: any, options: HierarchyOption): any[] {
   options = assign({} as HierarchyOption, DEFAULT_OPTIONS, options);
@@ -35,10 +41,12 @@ export function treemap(data: any, options: HierarchyOption): any[] {
     console.warn(e);
   }
 
+  const tileMethod = getTileMethod(options.tile, options.ratio);
+
   const partition = (data) =>
     d3Hierarchy
       .treemap()
-      .tile(d3Hierarchy[options.tile])
+      .tile(tileMethod)
       .size(options.size)
       .round(options.round)
       .padding(options.padding)